+++ src/main/java/SSVG/TPFinalDeliveryMascotero/Model/DetallePedidoEntity.java
@@ -7,6 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -23,7 +25,7 @@ public class DetallePedidoEntity {
     private Integer cantidad;
 
     @Column(nullable = false)
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
 
     private Double descuentoAplicado;
